Add custom HTML option for ads
This code adds the option to display custom HTML instead of the ads
being pulled in from Adsense. Custom HTML is boxes for each of the four
ad slots can be edited in the settings. It is currently just a text field
but we may want to edit it eventually to be a custom template that is a
textfield instead but this works.

The custom HTML that is typed inside the box is then rendered in the template
instead of the ad if it isn't empty. There are custom HTML boxes for both
the mobile and desktop views.